requirements.txt
requests>=2.31.0
streamlit>=1.28.0
pandas>=2.0.0
solders>=0.19.0
base58>=2.1.1
python-dotenv>=1.0.0
urllib3>=2.0.0
